[clang-tidy] Convert C casts to C++ ones
Found with google-readability-casting

+++ src/pager.cpp
@@ -421,7 +421,7 @@ bool pager_t::completion_try_print(size_t cols, const wcstring &prefix, const co
         this->available_term_height - 1 -
         (search_field_shown ? 1 : 0);  // we always subtract 1 to make room for a comment row
     if (!this->fully_disclosed) {
-        term_height = std::min(term_height, (size_t)PAGER_UNDISCLOSED_MAX_ROWS);
+        term_height = std::min(term_height, static_cast<size_t>(PAGER_UNDISCLOSED_MAX_ROWS));
     }
 
     size_t row_count = divide_round_up(lst.size(), cols);
@@ -494,7 +494,7 @@ bool pager_t::completion_try_print(size_t cols, const wcstring &prefix, const co
     assert(rendering->remaining_to_disclose != 1);
     if (rendering->remaining_to_disclose > 1) {
         progress_text = format_string(_(L"%lsand %lu more rows"), get_ellipsis_str(),
-                                      (unsigned long)rendering->remaining_to_disclose);
+                                      static_cast<unsigned long>(rendering->remaining_to_disclose));
     } else if (start_row > 0 || stop_row < row_count) {
         // We have a scrollable interface. The +1 here is because we are zero indexed, but want
         // to present things as 1-indexed. We do not add 1 to stop_row or row_count because
@@ -566,7 +566,7 @@ page_rendering_t pager_t::render() const {
             continue;
         }
 
-        rendering.cols = (size_t)cols;
+        rendering.cols = cols;
         rendering.rows = min_rows_required_for_cols;
         rendering.selected_completion_idx =
             this->visual_selected_completion_index(rendering.rows, rendering.cols);
@@ -659,7 +659,7 @@ bool pager_t::select_next_completion_in_direction(selection_motion_t direction,
         // Cardinal directions. We have a completion index; we wish to compute its row and column.
         size_t current_row = this->get_selected_row(rendering);
         size_t current_col = this->get_selected_column(rendering);
-        size_t page_height = std::max(rendering.term_height - 1, (size_t)1);
+        size_t page_height = std::max(rendering.term_height - 1, static_cast<size_t>(1));
 
         switch (direction) {
             case selection_motion_t::page_north: {
